Add npm command for react install on windows/linux
The react native cli installation info ( https://facebook.github.io/react-native/docs/getting-started.html#the-react-native-cli ) is missing the npm command to install react native cli if the platform is Windows or Linux. The code block only shows up if MacOS is selected as the development OS, but the npm command is identical on all platforms. It also makes it confusing for windows users because this crucial step is missing.